+++ src/Parser/ObjectParser.php
@@ -38,6 +38,8 @@ use const PHP_EOL;
  * ObjectParser
  *
  * Recursive object/variable parser
+ * 
+ * TODO: version bump
  *
  * @version 1.0.1
  */
@@ -51,10 +53,25 @@ class ObjectParser {
     public static int $depth = 4;
 
     /**
+     * Max dump depth
+     *
+     * N.B.: does not effect `var_dump`
+     * @see \Inane\Dumper\Dumper::$useVarExport
+     */
+    public int $parseDepth {
+        get => isset($this->parseDepth) ? $this->parseDepth : static::$depth;
+        set(int $value) {
+            $this->parseDepth = $value;
+        }
+    }
+
+    /**
      * Object Parser: private constructor
      *  So instantiation can only be done internally.
      */
-    private function __construct() {
+    private function __construct(?int $parseDepth = null) {
+        if ($parseDepth !== null)
+            $this->parseDepth = $parseDepth;
     }
 
     /**
@@ -65,16 +82,16 @@ class ObjectParser {
      *
      * @return string array as string
      */
-    private static function parseArray(array $array, int $level): string {
+    private function parseArray(array $array, int $level): string {
         $output = '';
 
-        if (static::$depth <= $level) $output .= '[...]';
+        if ($this->parseDepth <= $level) $output .= '[...]';
         else if (empty($array)) $output .= '[]';
         else {
             $keys = array_keys($array);
             $spaces = str_repeat(' ', $level * 4);
             $output .= '[';
-            foreach ($keys as $key) $output .= PHP_EOL . "{$spaces}    [$key] => " . self::parseVariable($array[$key], $level + 1). ',';
+            foreach ($keys as $key) $output .= PHP_EOL . "{$spaces}    [$key] => " . $this->parseVariable($array[$key], $level + 1). ',';
             $output .= PHP_EOL . "{$spaces}]";
         }
 
@@ -90,12 +107,12 @@ class ObjectParser {
      *
      * @return string object as string
      */
-    private static function parseObject(mixed $object, int $level, array &$cache): string {
+    private function parseObject(mixed $object, int $level, array &$cache): string {
         $output = '';
         $className = get_class($object);
 
         if (($id = array_search($object, $cache, true)) !== false) $output .= "{$className}#" . (++$id) . '(...)';
-        else if (static::$depth <= $level) $output .= "{$className}(...)";
+        else if ($this->parseDepth <= $level) $output .= "{$className}(...)";
         else {
             $id = array_push($cache, $object);
             $members = (array)$object;
@@ -105,7 +122,7 @@ class ObjectParser {
 
             foreach ($keys as $key) {
                 $keyDisplay = strtr(trim("$key"), ["\0" => ':']);
-                $output .= PHP_EOL . "{$spaces}    [$keyDisplay] => " . self::parseVariable($members[$key], $level + 1, $cache);
+                $output .= PHP_EOL . "{$spaces}    [$keyDisplay] => " . $this->parseVariable($members[$key], $level + 1, $cache);
             }
             $output .= PHP_EOL . "{$spaces}}";
         }
@@ -122,7 +139,7 @@ class ObjectParser {
      *
      * @return string dump string
      */
-    private static function parseVariable(mixed $var, int $level = 0, array &$cache = []): string {
+    private function parseVariable(mixed $var, int $level = 0, array &$cache = []): string {
         return match (gettype($var)) {
             'boolean' => $var ? 'true' : 'false',
             'integer', 'double' => "$var",
@@ -130,10 +147,22 @@ class ObjectParser {
             'resource' => '{resource}',
             'NULL' => 'null',
             'unknown type' => '{unknown}',
-            'array' => static::parseArray($var, $level),
-            'object' => static::parseObject($var, $level, $cache),
+            'array' => $this->parseArray($var, $level),
+            'object' => $this->parseObject($var, $level, $cache),
             default => '{unhandled}',
         };
+    }
+
+    /**
+     * Parses the given object and returns its string representation.
+     *
+     * @param mixed $object The object to be parsed.
+     * @param int|null $parseDepth Optional. The maximum depth to parse nested objects. If null, uses default depth.
+     * 
+     * @return string The string representation of the parsed object.
+     */
+    public static function parse(mixed $object, ?int $parseDepth = null): string {
+        return new self($parseDepth)->dump($object);
     }
 
     /**
@@ -143,7 +172,7 @@ class ObjectParser {
      *
      * @return string $object as string
      */
-    public static function parse(mixed $object): string {
-        return static::parseVariable($object);
+    public function dump(mixed $object): string {
+        return $this->parseVariable($object);
     }
 }